+++ lib_cuda/ray.h
@@ -6,7 +6,7 @@ class ray
 {
 public:
     __device__ ray() {}
-    __device__ ray(const vec3& a, const vec3& b, double t)
+    __device__ ray(const vec3& a, const vec3& b, float t)
     {
         A = a;
         B = b;
@@ -14,12 +14,12 @@ public:
     }
     __device__ vec3 origin() const { return A; }
     __device__ vec3 direction() const { return B; }
-    __device__ double time() const { return _time; }
-    __device__ vec3 point_at_parameter(double t) const { return A + t*B; }
+    __device__ float time() const { return _time; }
+    __device__ vec3 point_at_parameter(float t) const { return A + t*B; }
 
     vec3 A;
     vec3 B;
-    double _time;
+    float _time;
 };
 
 #endif
